Adds a helper function to create a config.
I need to create a config file in the CLI and during testing.

I was trying to keep a test config version up todate, but whenever I add a new param I had to recreate it which is a pain. I will move testing over to an autogenerated config based on the current config data model.