fix(workflow/scripts/bpm2abf): change outputs to use NA instead of zerors
In issue #30 we found that the verifyIDintensity will skip markers that are marked with an abf of "NA". The legacy was setting abf to 0 for a number of reasons. These markers were still being used in contamination calculations which seems incorrect. We changed the logic to output NAs which are skipped. Also changed logic to make multiallelic SNPs be set to NA. This mirrors the legacy script and fits the model assumptions of biallelic loci.

Ref #30