feat: add cache mechanism to download the GRAF software and support data
GRAF is not in Bioconda. I am not sure the best way to distribute it b/c it has a bunch of data packaged with the binary. Here I hack together a temporary solution to download the binary to a cached location. I can then dynamically add this location to the path before running shell commands.