fix(workflow/scripts/bim_filter_vcf): add exception handling for when allele is 0
For GRAF, I am now wanting to run the CR2 files directly in this script because GRAF does not need MAF or LD pruning. I ran into an exception when trying to get the complement of an allele `0`. I just add some try/except logic to return 0 instead of trying to complement 0.