fix(workflow/scripts/bpm2abf): add logic to handle impossible variant position
In #45, we found examples where variants, in the BPM file generated by Illumina, had as position of 0. This is not possible and was thowing an error when querying for the SNP in the VCF file. We already had defensive logic to handle chromosomes not in the VCF. Here I add logic to protect against impossible locations (less than 1 or greater than chromosome length).

Ref #45