Add Biowulf support (#65)
* feat(submission): add basic biowulf support

Adds basic biowulf cluster profile. Still need to test on biowulf.

* feat(cli): add confirmation message after job submit

Closes #66

* refactor(submission): move group componenets

Instead of defining groups on the command line for group-components. I
added group names to each of the `per_sample` rules.

* refactor: remove temp status for per_sample_* rules

I am still debugging the cluster and the per_sample rules are a pain to run again. I removed their temp status for now until I get things running smoothly.

* refactor: change plink resources

The plink resources were way too high for the actual need. We will need to benchmark on a large project to make final adjustments.

* feat: add resources for running rules locally

When we have <1k samples it does not make sense to submit a lot of rules
as a job. Add logic to cgr-submit to dynamically add more resources to
the main process for running a lot of rules locally.

* test(ci): disable one test when running with GitHub Actions