refactor(population_level): change checkpoint rule
I have been having issues with the workflow running to completation on the cluster. This seems to be a possible bug in checkpoints or maybe how I am using checkpoints. I decided to change my checkpoints to be closer aligned to how the documentation describes their use. Instead of using the qc_summary table as a checkpoint I am making the subject/control lists checkpoints. This will actually generate new files that are needed in the aggregation rule. Not sure if this will work but it is more similar to how snakemake does things.