fix(sample_qc_table): adds back the call rate filter flags
Here I change the flags up a bit, renaming the columns `cr1_filtered` and `cr2_filtered`. I also propegating missing data into the flags. So instead if a sample is missing, instead of saying the sample was removed by call rate it will just be missing. My hope is to make tracking where samples are filtered a little clearer.