Move workflow away from sample manifest to sample sheet (#134)
One major limitation was the use of our internal LIMS sample manifest file throughout the workflow. Here I refactor the entire workflow to (1) convert the LIMS sample manifest into `cgr_sample_sheet.csv` during pre-flight checks, (2) add logic so the user can start with just a CSV file instead of the sample manifest, (3) adjust the workflow to use the `cgr_sample_sheet.csv` instead of the LIMS sample manifest. 

The only required column is `Sample_ID`. In the config, the user also has to define `subject_id_column`, `expected_sex_column`, and `case_control_column`.

This simplifies a lot of code because we move from user input into an expected format (i.e., `cgr_sample-sheet.csv`) before we ever run the workflow. This allows me to move a lot of logic into central locations, so we are not doing the same thing in multiple places. Unfortunately, this required the re-writing of a lot of tests and may lead to some unforeseen bugs. 

Also, it required me to touch a lot of files in this PR.