feat(cluster_profiles): dynamically update resrouce requirements
This is a bit of a hack. For group jobs, snakemake sums all of their resources. So if I grouped a 1000 jobs together I can end up with asking for 1000 threads which is not what I want. Here I add logic back to dynamically set threads, mem, and time to be more appropriate for a individual job.