fix(workflow): skip SNP id update if not an rsID
Before running GRAF, we need to update SNP IDs to use rsID. The hg38 version of
the 1kg VCF is not well annotated. I ended updating SNP IDs that were rsID to
position based IDs. I added logic to skip the update if the 1kg ID is not an
rsID. This is mostly important for hg38's 1kg VCF which is not annotated.