Update the location where the real test data is stored (#201)
I recently move the location of the real test data repository. Here I update the the location for sync. I also wanted to added better configuration support for hg37 and hg38. I added a config option `genome_build` and a command line option to `cgr config`. This allows for automatic population of the correct reference files on CGEMs when using the different references.

# Closes 195