Remove sequential job grouping and run more commands locally
Using the group directive never worked correctly. When a grouped job finished the timestamps of all outputs were updated. So given a group jobs with `rule1 -> rule2 -> rule3` The timestamps of rule1 could end up newer than rule3. This caused the workflow to re-run all of the steps if the workflow is triggered again. This normally would not be a problem, but when used with checkpoints this caused parts of the workflow to be re-run 2 or 3 times. I did away with the `group` directive and moved a lot of the rules to run with the local snakemake process. I also updated the submit cli so that you can change the local resources.